Add files via upload
The code successfully trains a convolutional neural network model for handwritten number recognition using the MNIST dataset. The model achieves a certain level of accuracy on the test set and provides insights into its performance through visualizations such as accuracy and loss plots, a confusion matrix, and a classification report.